Rename NBuildStepsRunner to BuildStepsRunner and remove original class (#1124)
